+++ src/controllers/GameHistory.js
@@ -82,6 +82,6 @@ exports.getLastProtectedPlayer = async gameId => {
 
 exports.getLastVotePlay = gameId => this.findOne({ gameId, "play.action": "vote" }, null, { sort: { createdAt: -1 } });
 
-exports.getLastSistersPlay = gameId => this.findOne({ gameId, "play.source": "two-sisters" }, null, { sort: { createdAt: -1 } });
+exports.getLastSistersPlay = gameId => this.findOne({ gameId, "play.source.name": "two-sisters" }, null, { sort: { createdAt: -1 } });
 
-exports.getLastBrothersPlay = gameId => this.findOne({ gameId, "play.source": "three-brothers" }, null, { sort: { createdAt: -1 } });
+exports.getLastBrothersPlay = gameId => this.findOne({ gameId, "play.source.name": "three-brothers" }, null, { sort: { createdAt: -1 } });