+++ src/db/schemas/GameHistory.js
@@ -11,9 +11,15 @@ const playSchema = new Schema({
         required: true,
     },
     source: {
-        type: String,
-        enum: waitingForPossibilities,
-        required: true,
+        name: {
+            type: String,
+            enum: waitingForPossibilities,
+            required: true,
+        },
+        players: {
+            type: [Player],
+            required: true,
+        },
     },
     targets: {
         type: [
@@ -57,28 +63,6 @@ const playSchema = new Schema({
     versionKey: false,
 });
 
-/*
- * const eventSchema = new Schema({
- *     type: {
- *         type: String,
- *         required: true,
- *     },
- *     source: {
- *         type: String,
- *         enum: waitingForPossibilities,
- *         required: true,
- *     },
- *     targets: {
- *         type: [Player],
- *         default: undefined,
- *     },
- * }, {
- *     _id: false,
- *     timestamps: false,
- *     versionKey: false,
- * });
- */
-
 const gameHistorySchema = new Schema({
     gameId: {
         type: Schema.Types.ObjectId,
@@ -104,12 +88,6 @@ const gameHistorySchema = new Schema({
         type: playSchema,
         required: false,
     },
-    /*
-     * event: {
-     *     type: eventSchema,
-     *     required: false,
-     * },
-     */
 }, {
     timestamps: true,
     versionKey: false,